# lib/api-zod/src/generated/types/bleCommandRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Even Realities Dev Studio API — BLE simulator, display preview, and feature workbench
 * OpenAPI spec version: 0.2.0
 */
import type { BleCommandRequestCommandType } from "./bleCommandRequestCommandType";
import type { BleCommandRequestModel } from "./bleCommandRequestModel";
import type { BleCommandRequestPayload } from "./bleCommandRequestPayload";
import type { BleCommandRequestTarget } from "./bleCommandRequestTarget";

export interface BleCommandRequest {
  /** Target glasses model */
  model: BleCommandRequestModel;
  /** Which arm to send to */
  target: BleCommandRequestTarget;
  /** Type of command */
  commandType: BleCommandRequestCommandType;
  /** Command-specific payload */
  payload?: BleCommandRequestPayload;
  /** Raw hex bytes for custom commands (e.g. "0x0E 0x01") */
  rawHex?: string;
}