# lib/api-zod/src/generated/types/bleCommandRequestModel.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Even Realities Dev Studio API — BLE simulator, display preview, and feature workbench
 * OpenAPI spec version: 0.2.0
 */

/**
 * Target glasses model
 */
export type BleCommandRequestModel =
  (typeof BleCommandRequestModel)[keyof typeof BleCommandRequestModel];

export const BleCommandRequestModel = {
  G1: "G1",
  G2: "G2",
} as const;