# lib/api-zod/src/generated/types/blePacket.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Even Realities Dev Studio API — BLE simulator, display preview, and feature workbench
 * OpenAPI spec version: 0.2.0
 */
import type { BlePacketSide } from "./blePacketSide";

export interface BlePacket {
  index: number;
  hexBytes: string;
  sizeBytes: number;
  side: BlePacketSide;
}