# lib/api-zod/src/generated/types/textLayoutRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Even Realities Dev Studio API — BLE simulator, display preview, and feature workbench
 * OpenAPI spec version: 0.2.0
 */
import type { TextLayoutRequestModel } from "./textLayoutRequestModel";

export interface TextLayoutRequest {
  model: TextLayoutRequestModel;
  text: string;
  linesPerScreen?: number;
  /** Max display width in pixels (G1=488, G2=640) */
  maxWidthPx?: number;
  fontSizePx?: number;
}