# lib/api-zod/src/generated/types/textLayoutResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Even Realities Dev Studio API — BLE simulator, display preview, and feature workbench
 * OpenAPI spec version: 0.2.0
 */
import type { TextPage } from "./textPage";

export interface TextLayoutResponse {
  model: string;
  totalLines: number;
  totalPages: number;
  totalPackets: number;
  pages: TextPage[];
  displayWidthPx: number;
  linesPerScreen: number;
}